Copy the agent init script to the right location
Now, we judge the python verison to put stetho-agent script
to the right place. In CentOS 6.5, the python verison is '2.6'
and in CentOS 7, the python version is '2.7'. So in CentOS 6.5,
we should put the agent init script to the "/etc/init.d" and in
CentOS 7, we should put the agent init scriot to the "/etc/systemd/system/"