Restructure the docs for steth
Reformat the README.md into README.rst. Reschedule the
directory of the docs. Add code blocks to emphasize the
commands in the quickstart.rst. Seperate the multi-node-
architecture section out of the README.rst, and put it
in the multi-node-architecture.rst.

Note: The multi-node-architecture.rst is not indicated in
the README.rst.

Change-Id: I05407cf75e4f140158c690d96a9ded5f39cf6bc8
Closes-Bug: #1569139